Finishing implementation + testing of crossing%
A crossing has a id, connection and segment-list. These can also be retrieved by setters. You can change the state of the crossing with set-state!. Changing the state of the crossing takes time, so a seperate thread is launched, where every second, the state of the crossing is checked

User story #1, tasks #10 and #11